+++ src/CameraFollower.cpp
@@ -3,7 +3,7 @@
 
 CameraFollower::CameraFollower(GameObject& go) : Component(go) {}
 
-void CameraFollower::Update(float dt) {
+void CameraFollower::Update(double dt) {
     associated.box.x = -Camera::pos.x;
     associated.box.y = -Camera::pos.y;
 }
